changed the state machine so that a short button press can also be used to set unique addresses, if enabled you do need to pick an index first. The blinking LED should be clear in the matter.
Also the states of picking a signal type or selecting a unique address will fall back to getIndex instead of idle.
This allow you to configure the decoder faster. Pressing the button while not in idle, will abort any action.

Also swapped the 8th and 9th german signals and double red and red + shunt was swapped to correspond with documentation.
Website needs altering

Added state machine in yEd to match the code. And I updated the flow chart, though it may be revised for clarity's sake.